+++ test/libsolidity/SolidityNameAndTypeResolution.cpp
@@ -6428,34 +6428,34 @@ BOOST_AUTO_TEST_CASE(function_types_sig)
 	char const* text = R"(
 		contract C {
 			function f() returns (bytes4) {
-				return f.sig;
+				return f.selector;
 			}
 		}
 	)";
-	CHECK_ERROR(text, TypeError, "Member \"sig\" not found");
+	CHECK_ERROR(text, TypeError, "Member \"selector\" not found");
 	text = R"(
 		contract C {
 			function g() internal {
 			}
 			function f() returns (bytes4) {
-				return g.sig;
+				return g.selector;
 			}
 		}
 	)";
-	CHECK_ERROR(text, TypeError, "Member \"sig\" not found");
+	CHECK_ERROR(text, TypeError, "Member \"selector\" not found");
 	text = R"(
 		contract C {
 			function f() returns (bytes4) {
 				function () g;
-				return g.sig;
+				return g.selector;
 			}
 		}
 	)";
-	CHECK_ERROR(text, TypeError, "Member \"sig\" not found");
+	CHECK_ERROR(text, TypeError, "Member \"selector\" not found");
 	text = R"(
 		contract C {
 			function f() returns (bytes4) {
-				return this.f.sig;
+				return this.f.selector;
 			}
 		}
 	)";
@@ -6463,7 +6463,7 @@ BOOST_AUTO_TEST_CASE(function_types_sig)
 	text = R"(
 		contract C {
 			function f() external returns (bytes4) {
-				return this.f.sig;
+				return this.f.selector;
 			}
 		}
 	)";
@@ -6474,22 +6474,22 @@ BOOST_AUTO_TEST_CASE(function_types_sig)
 			}
 			function f() external returns (bytes4) {
 				var g = this.h;
-				return g.sig;
+				return g.selector;
 			}
 		}
 	)";
-	CHECK_ERROR(text, TypeError, "Member \"sig\" not found");
+	CHECK_ERROR(text, TypeError, "Member \"selector\" not found");
 	text = R"(
 		contract C {
 			function h() external {
 			}
 			function f() external returns (bytes4) {
 				function () external g = this.h;
-				return g.sig;
+				return g.selector;
 			}
 		}
 	)";
-	CHECK_ERROR(text, TypeError, "Member \"sig\" not found");
+	CHECK_ERROR(text, TypeError, "Member \"selector\" not found");
 }
 
 BOOST_AUTO_TEST_CASE(using_this_in_constructor)